requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pandas
scikit-learn
xgboost
lightgbm
optuna
joblib
matplotlib
gradio
python-jose
passlib[bcrypt]
pytest
ruff
bandit